Fix integer promotion mess
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@1801 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ generator/ppmpat.c
@@ -843,8 +843,8 @@ poles(pixel **     const pixels,
             dist1 = dist2 = (SQR((double)cols) + SQR((double)rows));
             for (i = 0; i < poleCt; ++i) {
                 double const newdist =
-                    (double)(col - xs[i]) * (col - xs[i]) +
-                    (double)(row - ys[i]) * (row - ys[i]);
+                    (double)((int)col - xs[i]) * ((int)col - xs[i]) +
+                    (double)((int)row - ys[i]) * ((int)row - ys[i]);
                 if (newdist < dist1) {
                     dist2  = dist1;
                     color2 = color1;
